Refactor for listeners and chat
