+++ bot_core/bot.py
@@ -3,19 +3,7 @@ from discord.ext import commands
 from jproperties import Properties
 from bot_core.commands.music import commands_music
 from bot_core.commands.chat import commands_chat
-
-
-# async def send_message(message: str, user_message: str, is_private: bool, is_custom: bool = None, motive: str = None):
-#     try:
-#         if is_custom:
-#             response = responses.handle_custom_message(motive)
-#             await message.channel.send(response)
-#         else:
-#             response = responses.handle_message(user_message)
-#             await message.author.send(response) if is_private else await message.channel.send(response)
-
-#     except Exception as e:
-#         print(f'Exception: {e}')
+# from bot_core.listeners.kick import listener_kick
 
 
 def run_discord_bot():
@@ -26,14 +14,15 @@ def run_discord_bot():
 
     token_key = configs.get('TOKEN').data
 
-    # Create client, bot and intents
+    # Create bot and intents
     intents = discord.Intents.all()
     bot = commands.Bot(command_prefix='.', intents=intents)
 
-    # @bot.event
-    # async def on_ready():
-    #     print(f'Boot {bot.user} is now running')
+    @bot.event
+    async def on_ready():
+        print(f'Boot {bot.user} is now running')
 
+    # listener_kick(bot)
     commands_chat(bot)
     commands_music(bot)
